add Table::insert

+++ src/data.cc
@@ -45,7 +45,7 @@ value::value(const value &v) : _type(v._type)
     }
 }
 
-value::value(double number) : _type(type_number)
+value::value(long long int number) : _type(type_number)
 {
     this->_storage.number_ = number;
 }
